Refactor card styles, inputs, and actions
