+++ ideas/button-tool-architecture.md
