+++ ideas/card-template-architecture.md
