+++ ideas/color-filter-architecture.md
