+++ ideas/consistent-color-style-names.md
