+++ ideas/control-content-stack.md
