+++ ideas/dynamic-tool-dimensions.md
