+++ ideas/entity-derived-state-architecture.md
