+++ ideas/entity-slots.md
