+++ ideas/fhs-input-boolean.md
