+++ ideas/gradient-runtime-color-architecture.md
