+++ ideas/home-assistant-actions-and-fhs-input-entities.md
