+++ ideas/horseshoe-colorstopgradient-arc-rendering.md
