+++ ideas/layout-item-colorstop-modes.md
