+++ ideas/layout-visibility-and-disabled.md
