+++ ideas/local-fhs-entity-click-routing.md
