+++ ideas/masks-clips-architecture-nl.md
