+++ ideas/masks-clips-architecture.md
