+++ ideas/ranking-stringstate-architecture.md
