+++ ideas/sparkline-automatic-bin-density.md
