+++ ideas/sparkline-chart-animations.md
