+++ ideas/sparkline-graph-tool-architecture-nl.md
