+++ ideas/sparkline-graph-tool-architecture.md
