+++ ideas/sparkline-history-lifecycle-architecture.md
