+++ ideas/sparkline-x-axis-label-sizing.md
