+++ ideas/state-bands-chart-architecture.md
