+++ ideas/text-tool-architecture.md
