+++ ideas/todo-sparkline.md
