+++ ideas/tool-hass-lifecycle-architecture.md
