+++ src/card-styles.js
@@ -1,0 +1,601 @@
+import { css } from "lit";
+
+/** Lit styles shared by every Flex Horseshoe Card instance. */
+const CardStyles = css`
+      :host {
+        cursor: default;
+      }
+
+      @media (print), (prefers-reduced-motion: reduce) {
+        .animated {
+          animation-duration: 1ms !important;
+          transition-duration: 1ms !important;
+          animation-iteration-count: 1 !important;
+        }
+      }
+
+      @keyframes zoomOut {
+        from {
+          opacity: 1;
+        }
+
+        50% {
+          opacity: 0;
+          transform: scale3d(0.3, 0.3, 0.3);
+        }
+
+        to {
+          opacity: 0;
+        }
+      }
+
+      @keyframes bounce {
+        from,
+        20%,
+        53%,
+        80%,
+        to {
+          animation-timing-function: cubic-bezier(0.215, 0.61, 0.355, 1);
+          transform: translate3d(0, 0, 0);
+        }
+
+        40%,
+        43% {
+          animation-timing-function: cubic-bezier(0.755, 0.05, 0.855, 0.06);
+          transform: translate3d(0, -30px, 0);
+        }
+
+        70% {
+          animation-timing-function: cubic-bezier(0.755, 0.05, 0.855, 0.06);
+          transform: translate3d(0, -15px, 0);
+        }
+
+        90% {
+          transform: translate3d(0, -4px, 0);
+        }
+      }
+
+      @keyframes flash {
+        from,
+        50%,
+        to {
+          opacity: 1;
+        }
+
+        25%,
+        75% {
+          opacity: 0;
+        }
+      }
+
+      @keyframes headShake {
+        0% {
+          transform: translateX(0);
+        }
+
+        6.5% {
+          transform: translateX(-6px) rotateY(-9deg);
+        }
+
+        18.5% {
+          transform: translateX(5px) rotateY(7deg);
+        }
+
+        31.5% {
+          transform: translateX(-3px) rotateY(-5deg);
+        }
+
+        43.5% {
+          transform: translateX(2px) rotateY(3deg);
+        }
+
+        50% {
+          transform: translateX(0);
+        }
+      }
+
+      @keyframes heartBeat {
+        0% {
+          transform: scale(1);
+        }
+
+        14% {
+          transform: scale(1.3);
+        }
+
+        28% {
+          transform: scale(1);
+        }
+
+        42% {
+          transform: scale(1.3);
+        }
+
+        70% {
+          transform: scale(1);
+        }
+      }
+
+      @keyframes jello {
+        from,
+        11.1%,
+        to {
+          transform: translate3d(0, 0, 0);
+        }
+
+        22.2% {
+          transform: skewX(-12.5deg) skewY(-12.5deg);
+        }
+
+        33.3% {
+          transform: skewX(6.25deg) skewY(6.25deg);
+        }
+
+        44.4% {
+          transform: skewX(-3.125deg) skewY(-3.125deg);
+        }
+
+        55.5% {
+          transform: skewX(1.5625deg) skewY(1.5625deg);
+        }
+
+        66.6% {
+          transform: skewX(-0.78125deg) skewY(-0.78125deg);
+        }
+
+        77.7% {
+          transform: skewX(0.390625deg) skewY(0.390625deg);
+        }
+
+        88.8% {
+          transform: skewX(-0.1953125deg) skewY(-0.1953125deg);
+        }
+      }
+
+      @keyframes pulse {
+        from {
+          transform: scale3d(1, 1, 1);
+        }
+
+        50% {
+          transform: scale3d(1.05, 1.05, 1.05);
+        }
+
+        to {
+          transform: scale3d(1, 1, 1);
+        }
+      }
+
+      @keyframes rubberBand {
+        from {
+          transform: scale3d(1, 1, 1);
+        }
+
+        30% {
+          transform: scale3d(1.25, 0.75, 1);
+        }
+
+        40% {
+          transform: scale3d(0.75, 1.25, 1);
+        }
+
+        50% {
+          transform: scale3d(1.15, 0.85, 1);
+        }
+
+        65% {
+          transform: scale3d(0.95, 1.05, 1);
+        }
+
+        75% {
+          transform: scale3d(1.05, 0.95, 1);
+        }
+
+        to {
+          transform: scale3d(1, 1, 1);
+        }
+      }
+
+      @keyframes shake {
+        from,
+        to {
+          transform: translate3d(0, 0, 0);
+        }
+
+        10%,
+        30%,
+        50%,
+        70%,
+        90% {
+          transform: translate3d(-10px, 0, 0);
+        }
+
+        20%,
+        40%,
+        60%,
+        80% {
+          transform: translate3d(10px, 0, 0);
+        }
+      }
+
+      @keyframes swing {
+        20% {
+          transform: rotate3d(0, 0, 1, 15deg);
+        }
+
+        40% {
+          transform: rotate3d(0, 0, 1, -10deg);
+        }
+
+        60% {
+          transform: rotate3d(0, 0, 1, 5deg);
+        }
+
+        80% {
+          transform: rotate3d(0, 0, 1, -5deg);
+        }
+
+        to {
+          transform: rotate3d(0, 0, 1, 0deg);
+        }
+      }
+
+      @keyframes tada {
+        from {
+          transform: scale3d(1, 1, 1);
+        }
+        10%,
+        20% {
+          transform: scale3d(0.9, 0.9, 0.9) rotate3d(0, 0, 1, -3deg);
+        }
+        30%,
+        50%,
+        70%,
+        90% {
+          transform: scale3d(1.1, 1.1, 1.1) rotate3d(0, 0, 1, 3deg);
+        }
+        40%,
+        60%,
+        80% {
+          transform: scale3d(1.1, 1.1, 1.1) rotate3d(0, 0, 1, -3deg);
+        }
+        to {
+          transform: scale3d(1, 1, 1);
+        }
+      }
+
+      @keyframes wobble {
+        from {
+          transform: translate3d(0, 0, 0);
+        }
+        15% {
+          transform: translate3d(-25%, 0, 0) rotate3d(0, 0, 1, -5deg);
+        }
+        30% {
+          transform: translate3d(20%, 0, 0) rotate3d(0, 0, 1, 3deg);
+        }
+        45% {
+          transform: translate3d(-15%, 0, 0) rotate3d(0, 0, 1, -3deg);
+        }
+        60% {
+          transform: translate3d(10%, 0, 0) rotate3d(0, 0, 1, 2deg);
+        }
+        75% {
+          transform: translate3d(-5%, 0, 0) rotate3d(0, 0, 1, -1deg);
+        }
+        to {
+          transform: translate3d(0, 0, 0);
+        }
+      }
+
+      @media screen and (min-width: 467px) {
+        :host {
+          font-size: 12px;
+        }
+      }
+      @media screen and (max-width: 466px) {
+        :host {
+          font-size: 12px;
+        }
+      }
+
+      :host ha-card {
+        padding: 5px 5px 5px 5px;
+      }
+
+      :host([embedded]) {
+        display: block;
+        width: 100%;
+        height: 100%;
+      }
+
+      :host([embedded]) ha-card {
+        box-sizing: border-box;
+        width: 100%;
+        height: 100%;
+      }
+
+      :host([embedded]) .container {
+        width: 100%;
+        height: 100%;
+      }
+
+      :host([embedded]) .container > svg {
+        width: 100%;
+        height: 100%;
+      }
+
+      .container {
+        position: relative;
+        height: 100%;
+        display: flex;
+        flex-direction: column;
+      }
+
+      .fhs-child-card-layer {
+        position: absolute;
+        inset: 0;
+        pointer-events: none;
+      }
+
+      .sparkline-tooltip-layer {
+        position: absolute;
+        inset: 0;
+        pointer-events: none;
+      }
+
+      .sparkline-tooltip {
+        position: absolute;
+        z-index: 5;
+        pointer-events: none;
+        display: inline-block;
+        width: auto;
+        max-width: calc(100% - 24px);
+        padding: 0.2em 0.3em;
+        border-radius: 0.3em;
+        background: var(--card-background-color, var(--ha-card-background, rgba(32, 32, 32, 0.94)));
+        color: var(--primary-text-color);
+        box-shadow: 0 0.35em 0.9em rgba(0, 0, 0, 0.22);
+        border: 1px solid var(--divider-color);
+        font-size: var(--sparkline-tooltip-font-size, 0.5em);
+        -webkit-text-size-adjust: 100%;
+        text-size-adjust: 100%;
+        line-height: 1.15;
+        transform: translate(-50%, -100%);
+      }
+
+      .sparkline-tooltip__title {
+        font-weight: 600;
+        margin-bottom: 0.22em;
+        white-space: nowrap;
+      }
+
+      .sparkline-tooltip__row {
+        display: grid;
+        grid-template-columns: auto auto;
+        gap: 0.6em;
+        align-items: baseline;
+        white-space: nowrap;
+      }
+
+      .sparkline-tooltip__row + .sparkline-tooltip__row {
+        margin-top: 0.08em;
+      }
+
+      .fhs-child-card {
+        position: absolute;
+        pointer-events: auto;
+      }
+
+      .fhs-child-card > * {
+        display: block;
+        width: 100%;
+        height: 100%;
+      }
+
+      .fhs-child-card--frameless {
+        background: transparent;
+        border: 0;
+        box-shadow: none;
+      }
+
+      .labelContainer {
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 65%;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: flex-end;
+      }
+
+      .ellipsis {
+        text-overflow: ellipsis;
+        white-space: nowrap;
+        overflow: hidden;
+      }
+
+      .icon-svg-url.hidden {
+        display: none;
+      }
+
+      .state {
+        position: relative;
+        display: flex;
+        flex-wrap: wrap;
+        max-width: 100%;
+        min-width: 0px;
+      }
+
+      #label {
+        display: flex;
+        line-height: 1;
+      }
+
+      #label.bold {
+        font-weight: bold;
+      }
+
+      #label,
+      #name {
+        margin: 3% 0;
+      }
+
+      .text {
+        font-size: 100%;
+      }
+
+      #name {
+        font-size: 80%;
+        font-weight: 300;
+      }
+
+      .unit {
+        font-size: 65%;
+        font-weight: normal;
+        opacity: 0.6;
+        line-height: 2em;
+        vertical-align: bottom;
+        margin-left: 0.25rem;
+      }
+
+      .entity__area {
+        position: absolute;
+        top: 70%;
+        font-size: 120%;
+        opacity: 0.6;
+        display: flex;
+        line-height: 1;
+        align-items: center;
+        justify-content: center;
+        width: 100%;
+        height: 20%;
+        flex-direction: column;
+      }
+
+      .nam {
+        alignment-baseline: central;
+        fill: var(--primary-text-color);
+      }
+
+      .state__uom {
+        font-size: 20px;
+        opacity: 0.7;
+        margin: 0;
+        fill: var(--primary-text-color);
+      }
+
+      .state__value {
+        font-size: 3em;
+        opacity: 1;
+        fill: var(--primary-text-color);
+        text-anchor: middle;
+      }
+      .entity__name {
+        text-anchor: middle;
+        overflow: hidden;
+        opacity: 0.8;
+        fill: var(--primary-text-color);
+        font-size: 1.5em;
+        text-transform: uppercase;
+        letter-spacing: 0.1em;
+      }
+
+      .entity__area {
+        font-size: 12px;
+        opacity: 0.7;
+        overflow: hidden;
+        fill: var(--primary-text-color);
+        text-anchor: middle;
+        text-transform: uppercase;
+        letter-spacing: 0.1em;
+      }
+
+      .shadow {
+        font-size: 30px;
+        font-weight: 700;
+        text-anchor: middle;
+      }
+
+      .card--dropshadow-5 {
+        filter: drop-shadow(0 1px 0 #ccc) drop-shadow(0 2px 0 #c9c9c9) drop-shadow(0 3px 0 #bbb) drop-shadow(0 4px 0 #b9b9b9) drop-shadow(0 5px 0 #aaa) drop-shadow(0 6px 1px rgba(0, 0, 0, 0.1))
+          drop-shadow(0 0 5px rgba(0, 0, 0, 0.1)) drop-shadow(0 1px 3px rgba(0, 0, 0, 0.3)) drop-shadow(0 3px 5px rgba(0, 0, 0, 0.2)) drop-shadow(0 5px 10px rgba(0, 0, 0, 0.25))
+          drop-shadow(0 10px 10px rgba(0, 0, 0, 0.2)) drop-shadow(0 20px 20px rgba(0, 0, 0, 0.15));
+      }
+      .card--dropshadow-medium--opaque--sepia90 {
+        filter: drop-shadow(0em 0.05em 0px #b2a98f22) drop-shadow(0em 0.07em 0px #b2a98f55) drop-shadow(0em 0.1em 0px #b2a98f88) drop-shadow(0px 0.6em 0.9em rgba(0, 0, 0, 0.15))
+          drop-shadow(0px 1.2em 0.15em rgba(0, 0, 0, 0.1)) drop-shadow(0px 2.4em 2.5em rgba(0, 0, 0, 0.1)) sepia(90%);
+      }
+
+      .card--dropshadow-heavy--sepia90 {
+        filter: drop-shadow(0em 0.05em 0px #b2a98f22) drop-shadow(0em 0.07em 0px #b2a98f55) drop-shadow(0em 0.1em 0px #b2a98f88) drop-shadow(0px 0.3em 0.45em rgba(0, 0, 0, 0.5))
+          drop-shadow(0px 0.6em 0.07em rgba(0, 0, 0, 0.3)) drop-shadow(0px 1.2em 1.25em rgba(0, 0, 0, 1)) drop-shadow(0px 1.8em 1.6em rgba(0, 0, 0, 0.1)) drop-shadow(0px 2.4em 2em rgba(0, 0, 0, 0.1))
+          drop-shadow(0px 3em 2.5em rgba(0, 0, 0, 0.1)) sepia(90%);
+      }
+
+      .card--dropshadow-heavy {
+        filter: drop-shadow(0em 0.05em 0px #b2a98f22) drop-shadow(0em 0.07em 0px #b2a98f55) drop-shadow(0em 0.1em 0px #b2a98f88) drop-shadow(0px 0.3em 0.45em rgba(0, 0, 0, 0.5))
+          drop-shadow(0px 0.6em 0.07em rgba(0, 0, 0, 0.3)) drop-shadow(0px 1.2em 1.25em rgba(0, 0, 0, 1)) drop-shadow(0px 1.8em 1.6em rgba(0, 0, 0, 0.1)) drop-shadow(0px 2.4em 2em rgba(0, 0, 0, 0.1))
+          drop-shadow(0px 3em 2.5em rgba(0, 0, 0, 0.1));
+      }
+
+      .card--dropshadow-medium--sepia90 {
+        filter: drop-shadow(0em 0.05em 0px #b2a98f) drop-shadow(0em 0.15em 0px #b2a98f) drop-shadow(0em 0.15em 0px #b2a98f) drop-shadow(0px 0.6em 0.9em rgba(0, 0, 0, 0.15))
+          drop-shadow(0px 1.2em 0.15em rgba(0, 0, 0, 0.1)) drop-shadow(0px 2.4em 2.5em rgba(0, 0, 0, 0.1)) sepia(90%);
+      }
+
+      .card--dropshadow-medium {
+        filter: drop-shadow(0em 0.05em 0px #b2a98f) drop-shadow(0em 0.15em 0px #b2a98f) drop-shadow(0em 0.15em 0px #b2a98f) drop-shadow(0px 0.6em 0.9em rgba(0, 0, 0, 0.15))
+          drop-shadow(0px 1.2em 0.15em rgba(0, 0, 0, 0.1)) drop-shadow(0px 2.4em 2.5em rgba(0, 0, 0, 0.1));
+      }
+
+      .card--dropshadow-light--sepia90 {
+        filter: drop-shadow(0px 0.1em 0px #b2a98f) drop-shadow(0.1em 0.5em 0.2em rgba(0, 0, 0, 0.5)) sepia(90%);
+      }
+
+      .card--dropshadow-light {
+        filter: drop-shadow(0px 0.1em 0px #b2a98f) drop-shadow(0.1em 0.5em 0.2em rgba(0, 0, 0, 0.5));
+      }
+
+      .card--dropshadow-down-and-distant {
+        filter: drop-shadow(0px 0.05em 0px #b2a98f) drop-shadow(0px 14px 10px rgba(0, 0, 0, 0.15)) drop-shadow(0px 24px 2px rgba(0, 0, 0, 0.1)) drop-shadow(0px 34px 30px rgba(0, 0, 0, 0.1));
+      }
+      .card--filter-none {
+      }
+
+      .horseshoe__svg__group {
+        /*
+          * Was transform: translateY(15%).
+          * After fixing SVG viewBox/namespace parsing, this offset became visible
+          * and moved the horseshoe down.
+          * A nice 6 year old bug ;-)
+          */
+      }
+
+      .line__horizontal {
+        stroke: var(--primary-text-color);
+        opacity: 0.3;
+        stroke-width: 2;
+      }
+
+      .line__vertical {
+        stroke: var(--primary-text-color);
+        opacity: 0.3;
+        stroke-width: 2;
+      }
+
+      .svg__dot {
+        fill: var(--primary-text-color);
+        opacity: 0.5;
+        align-self: center;
+        transform-origin: 50% 50%;
+      }
+
+      .icon {
+        align: center;
+      }
+`;
+
+export default CardStyles;
